refactor: extract pack status logic to pkg/pack/status.go
Move core pack status functionality from commands layer to pack layer
to improve code organization and avoid import cycles. This follows the
established pattern used by other commands like init, adopt, and fill.

Changes:
- Create pkg/pack/status.go with GetStatus function and types
- Move status logic (handler status checking, pack status calculation)
- Keep display formatting and command orchestration in commands layer
- Convert between pack.StatusResult and types.DisplayPack for UI

The command now focuses on orchestration and display while the pack
package handles the core status logic. This separation makes the code
more modular and testable.
